% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_config.R
\name{extract_references_framework}
\alias{extract_references_framework}
\title{Extract join framework to put into gcode}
\usage{
extract_references_framework(verbose = T)
}
\arguments{
\item{alpha_sample}{Joining the alpha_sample parameters. A vector of integers, where identical integers indicate same the data axis to be joined.}

\item{beta_sample}{Joining the beta_sample parameters. A vector of integers, where identical integers indicate same the data axis to be joined.}

\item{code}{Joining the code parameters. A vector of integers, where identical integers indicate same the data axis to be joined.}
}
\description{
Join data to improve modelling capacity for similar axes
}
