% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_impute.R
\name{multiple_imputation}
\alias{multiple_imputation}
\title{Multiple Imputation via Generalised Canonical Procrustes}
\usage{
multiple_imputation(
  data_list,
  multi_impute = 5,
  config = gcode::extract_config(verbose = F),
  transfer = gcode::extract_transfer_framework(verbose = F),
  recover = gcode::extract_recovery_framework(verbose = F),
  join = gcode::extract_join_framework(verbose = F)
)
}
\arguments{
\item{config}{Configuration parameters (required, default provided)}

\item{transfer}{Transferring pre-trained model parameters (not required)}

\item{recover}{Important information used for prediction or imputation (not required)}

\item{join}{Join parameters that share the same axes to be similar (not required)}

\item{data}{List of data matrices of varying dimensionality. Attempts to find similarities among all datasets with a core structure.}
}
\description{
Multiple Imputation via Generalised Canonical Procrustes
}
