% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{transfer}
\alias{transfer}
\title{Transfer model from one to another}
\usage{
transfer(model_list, data_list, config = gcode::extract_config(verbose = F))
}
\arguments{
\item{model_list}{A list of gcode models}

\item{data_list}{A list of datasets (matrix or tensor etc.)}

\item{config}{Configuration parameters (required, default provided)}

\item{transfer}{Transferring pre-trained model parameters (not required)}

\item{join}{Join parameters that share the same axes to be similar (not required)}
}
\value{
All models to be learned or updated
}
\description{
Main function to transferring models
}
